// src/apispec/src/models/ProcessEvent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProcessEventType } from './ProcessEventType';
import {
    ProcessEventTypeFromJSON,
    ProcessEventTypeFromJSONTyped,
    ProcessEventTypeToJSON,
    ProcessEventTypeToJSONTyped,
} from './ProcessEventType';

/**
 * 
 * @export
 * @interface ProcessEvent
 */
export interface ProcessEvent {
    /**
     * 
     * @type {number}
     * @memberof ProcessEvent
     */
    seq: number;
    /**
     * 
     * @type {string}
     * @memberof ProcessEvent
     */
    eventId?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessEvent
     */
    processId: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessEvent
     */
    channel?: string;
    /**
     * 
     * @type {ProcessEventType}
     * @memberof ProcessEvent
     */
    type: ProcessEventType;
    /**
     * 
     * @type {Date}
     * @memberof ProcessEvent
     */
    timestamp: Date;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ProcessEvent
     */
    payload?: { [key: string]: any; };
}



/**
 * Check if a given object implements the ProcessEvent interface.
 */
export function instanceOfProcessEvent(value: object): value is ProcessEvent {
    if (!('seq' in value) || value['seq'] === undefined) return false;
    if (!('processId' in value) || value['processId'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('timestamp' in value) || value['timestamp'] === undefined) return false;
    return true;
}

export function ProcessEventFromJSON(json: any): ProcessEvent {
    return ProcessEventFromJSONTyped(json, false);
}

export function ProcessEventFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessEvent {
    if (json == null) {
        return json;
    }
    return {
        
        'seq': json['seq'],
        'eventId': json['event_id'] == null ? undefined : json['event_id'],
        'processId': json['process_id'],
        'channel': json['channel'] == null ? undefined : json['channel'],
        'type': ProcessEventTypeFromJSON(json['type']),
        'timestamp': (new Date(json['timestamp'])),
        'payload': json['payload'] == null ? undefined : json['payload'],
    };
}

export function ProcessEventToJSON(json: any): ProcessEvent {
    return ProcessEventToJSONTyped(json, false);
}

export function ProcessEventToJSONTyped(value?: ProcessEvent | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'seq': value['seq'],
        'event_id': value['eventId'],
        'process_id': value['processId'],
        'channel': value['channel'],
        'type': ProcessEventTypeToJSON(value['type']),
        'timestamp': value['timestamp'].toISOString(),
        'payload': value['payload'],
    };
}