// src/apispec/src/models/ProcessInputEvent.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProcessEventType } from './ProcessEventType';
import {
    ProcessEventTypeFromJSON,
    ProcessEventTypeFromJSONTyped,
    ProcessEventTypeToJSON,
    ProcessEventTypeToJSONTyped,
} from './ProcessEventType';

/**
 * 
 * @export
 * @interface ProcessInputEvent
 */
export interface ProcessInputEvent {
    /**
     * Idempotency key for retries.
     * @type {string}
     * @memberof ProcessInputEvent
     */
    eventId: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessInputEvent
     */
    channel: string;
    /**
     * 
     * @type {ProcessEventType}
     * @memberof ProcessInputEvent
     */
    type: ProcessEventType;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ProcessInputEvent
     */
    payload?: { [key: string]: any; };
}



/**
 * Check if a given object implements the ProcessInputEvent interface.
 */
export function instanceOfProcessInputEvent(value: object): value is ProcessInputEvent {
    if (!('eventId' in value) || value['eventId'] === undefined) return false;
    if (!('channel' in value) || value['channel'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function ProcessInputEventFromJSON(json: any): ProcessInputEvent {
    return ProcessInputEventFromJSONTyped(json, false);
}

export function ProcessInputEventFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessInputEvent {
    if (json == null) {
        return json;
    }
    return {
        
        'eventId': json['event_id'],
        'channel': json['channel'],
        'type': ProcessEventTypeFromJSON(json['type']),
        'payload': json['payload'] == null ? undefined : json['payload'],
    };
}

export function ProcessInputEventToJSON(json: any): ProcessInputEvent {
    return ProcessInputEventToJSONTyped(json, false);
}

export function ProcessInputEventToJSONTyped(value?: ProcessInputEvent | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'event_id': value['eventId'],
        'channel': value['channel'],
        'type': ProcessEventTypeToJSON(value['type']),
        'payload': value['payload'],
    };
}