// src/apispec/src/models/SignalContextRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global directory.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SignalContextRequest
 */
export interface SignalContextRequest {
    /**
     * 
     * @type {string}
     * @memberof SignalContextRequest
     */
    signal: string;
}

/**
 * Check if a given object implements the SignalContextRequest interface.
 */
export function instanceOfSignalContextRequest(value: object): value is SignalContextRequest {
    if (!('signal' in value) || value['signal'] === undefined) return false;
    return true;
}

export function SignalContextRequestFromJSON(json: any): SignalContextRequest {
    return SignalContextRequestFromJSONTyped(json, false);
}

export function SignalContextRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SignalContextRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'signal': json['signal'],
    };
}

export function SignalContextRequestToJSON(json: any): SignalContextRequest {
    return SignalContextRequestToJSONTyped(json, false);
}

export function SignalContextRequestToJSONTyped(value?: SignalContextRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'signal': value['signal'],
    };
}